ai/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
pillow>=8.3.1
matplotlib>=3.4.3
numpy>=1.21.2
opencv-python>=4.5.3
flask>=2.0.1
scikit-learn>=0.24.2
tqdm>=4.62.2
faiss-cpu>=1.7.1 